+++ src/utils/storage/index.ts
@@ -1,4 +1,4 @@
-import { LocalStorageManager } from './localStorage'
+import { LocalStorageManager } from './local-storage'
 
 export const AUTH_KEY = '@@auth-token'
 export const authTokenStorage = new LocalStorageManager<string>(AUTH_KEY)
